Feature to create, update and delete posts

+++ flask_blog/routes.py
@@ -1,32 +1,17 @@
 import os
 import secrets
 from PIL import Image
-from flask import url_for, render_template, flash, redirect, request
+from flask import url_for, render_template, flash, redirect, request, abort
 from flask_login import login_user, current_user, logout_user, login_required
-from .forms import RegistrationForm, LoginForm, UpdateAccountForm
+from .forms import RegistrationForm, LoginForm, UpdateAccountForm, PostForm
 from .models import User, Post
 from flask_blog import app, db, bcrypt
-
-# Test data
-posts = [
-    {
-        "author": "Davi Feliciano",
-        "title": "First Post!",
-        "content": "This is the first post on the blog!",
-        "date": "Epoch",
-    },
-    {
-        "author": "Victor Rebeque",
-        "title": "Second Post!",
-        "content": "This is the second post on the blog!",
-        "date": "Epoch Plus One",
-    },
-]
 
 
 @app.route("/")
 @app.route("/home")
 def homepage():
+    posts = Post.query.all()
     return render_template("homepage.html", title="Homepage", posts=posts)
 
 
@@ -84,7 +69,8 @@ def save_picture(form_picture):
     old_picture_path = os.path.join(
         app.root_path, "static/profile_pics", current_user.image_file
     )
-    os.remove(old_picture_path)
+    if current_user.image_file != "default.jpg":
+        os.remove(old_picture_path)
 
     random_hex = secrets.token_hex(8)
     _, file_ext = os.path.splitext(form_picture.filename)
@@ -119,3 +105,59 @@ def account():
     return render_template(
         "account.html", title="Account", image_file=image_file, form=form
     )
+
+
+@app.route("/post/new", methods=("GET", "POST"))
+@login_required
+def new_post():
+    form = PostForm()
+    if form.validate_on_submit():
+        post = Post(
+            title=form.title.data, content=form.content.data, author=current_user
+        )
+        db.session.add(post)
+        db.session.commit()
+        flash("Your post has been created!", "success")
+        return redirect(url_for("homepage"))
+    return render_template(
+        "create_post.html", title="New Post", form=form, legend="New Post"
+    )
+
+
+@app.route("/post/<int:post_id>")
+def post(post_id):
+    post = Post.query.get_or_404(post_id)
+    return render_template("post.html", title=post.title, post=post)
+
+
+@app.route("/post/<int:post_id>/update", methods=("GET", "POST"))
+@login_required
+def update_post(post_id):
+    post = Post.query.get_or_404(post_id)
+    if post.author != current_user:
+        abort(403)
+    form = PostForm()
+    form.title.data = post.title
+    form.content.data = post.content
+    if form.validate_on_submit():
+        db.session.commit()
+        flash("Your post has been updated!", "success")
+        return redirect(url_for("post", post_id=post.id))
+    elif request.method == "GET":
+        post.title = form.title.data
+        post.content = form.content.data
+    return render_template(
+        "create_post.html", title="New Post", form=form, legend="Update Post"
+    )
+
+
+@app.route("/post/<int:post_id>/delete", methods=["POST"])
+@login_required
+def delete_post(post_id):
+    post = Post.query.get_or_404(post_id)
+    if post.author != current_user:
+        abort(403)
+    db.session.delete(post)
+    db.session.commit()
+    flash("Your post has been deleted!", "success")
+    return redirect(url_for("homepage"))
